requirements.txt
imageio
ipdb
parse
kornia==0.4.1
lmdb==0.98
matplotlib==3.1.3
numpy==1.17.2
opencv-python==4.4.0.42
plyfile==0.7.1
pytorch-ignite==0.2.0
pytorch-lightning==0.7.5
scikit-image==0.18.1
scikit-learn==0.20.1
scipy==1.4.1
tqdm==4.56.0
visdom==0.1.8.9
